Changed names and checked pagination logic
