+++ src/controllers/users/get-like-email.js
@@ -1,6 +1,6 @@
 "use strict";
 //Services
-const { getLikeEmail } = require("../../services/users/getLikeEmail");
+const { getLikeEmail } = require("../../services/users/get-like-email");
 //Enums
 const { statusCode } = require("../../enums/http/statusCode");
 const { value } = require("../../enums/general/value");
@@ -14,6 +14,7 @@ const { validateAuthHeaders } = require("../../helpers/auth/headers");
 const {
   validatePathParameters,
 } = require("../../helpers/http/queryStringParams");
+const { checkOrderBy, checkOrderAt } = require("../../helpers/pagination/users/order");
 //Const/Vars
 let userList;
 let email;
@@ -26,7 +27,9 @@ let pageSizeNro;
 let pageNro;
 let msg;
 let code;
-const orderBy = [["id", "ASC"]];
+let orderAt;
+let orderBy;
+let order;
 
 /**
  * @description get all paged users whose email matches the passed as parameter
@@ -38,7 +41,9 @@ module.exports.handler = async (event) => {
     userList = value.IS_NULL;
     email = value.IS_NULL;
     pageSizeNro = 5;
-    pageNro = value.IS_ZERO_NUMBER;
+    pageNro = 0;
+    orderBy = "id";
+    orderAt = "ASC";
     msg = value.IS_NULL;
     code = value.IS_NULL;
 
@@ -85,11 +90,43 @@ module.exports.handler = async (event) => {
     if (queryStrParams != value.IS_NULL) {
       pageSizeNro = parseInt(await event.queryStringParameters.limit);
       pageNro = parseInt(await event.queryStringParameters.page);
+      pageNro = event.queryStringParameters.page
+      ? parseInt(await event.queryStringParameters.page)
+      : pageNro;
+    orderBy = event.queryStringParameters.orderBy
+      ? event.queryStringParameters.orderBy
+      : orderBy;
+    orderAt = event.queryStringParameters.orderAt
+      ? event.queryStringParameters.orderAt
+      : orderAt;
     }
+
+    
+    orderBy = await checkOrderBy(orderBy);
+
+    if(orderBy == (null || undefined)){
+      return await requestResult(
+        statusCode.BAD_REQUEST,
+        "It is not possible to apply sorting based on the requested orderBy value. Invalid field",
+        event
+      );
+    }
+
+    orderAt = await checkOrderAt(orderAt);
+
+    if(orderAt == (undefined || null)){
+      return await requestResult(
+        statusCode.BAD_REQUEST,
+        "It is not possible to apply sorting based on the requested orderAt value. Invalid field",
+        event
+      );
+    }
+    
+    order = [[orderBy, orderAt]];
     //-- end with pagination  ---
 
     //-- start with db query  ---
-    userList = await getLikeEmail(email, pageSizeNro, pageNro, orderBy);
+    userList = await getLikeEmail(email, pageSizeNro, pageNro, order);
 
     switch (userList) {
       case statusName.CONNECTION_REFUSED:
