+++ src/services/users/get-like-email.js
@@ -11,21 +11,16 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose identification type matches the passed as parameter
- * @param {String} identificationType String type
+ * @description get all paged users whose email matches the passed as parameter
+ * @param {String} email String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
- * @param {Object} orderBy Array Object type
+ * @param {Object} order Array Object type
  * @returns a list of paginated users
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeIdentificationType = async function (
-  identificationType,
-  pageSizeNro,
-  pageNro,
-  orderBy
-) {
+const getLikeEmail = async function (email, pageSizeNro, pageNro, order) {
   try {
     usersList = null;
     msg = null;
@@ -39,19 +34,19 @@ const getLikeIdentificationType = async function (
           ],
         },
         where: {
-          identification_type: {
-            [Op.like]: `%${identificationType}%`, //containing what is entered, less strictmatch
+          email: {
+            [Op.like]: `%${email}%`, //containing what is entered, less strictmatch
           },
         },
         limit: pageSizeNro,
         offset: pageNro,
-        order: orderBy,
+        order: order,
       })
         .then((users) => {
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeIdentificationType User model. Caused by ${error}`;
+          msg = `Error in getLikeEmail User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -59,14 +54,15 @@ const getLikeIdentificationType = async function (
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeIdentificationType function. Caused by ${error}`;
+    msg = `Error in getLikeEmail function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
+
   console.log(usersList);
   return usersList;
 };
 
 module.exports = {
-  getLikeIdentificationType,
+  getLikeEmail,
 };
