+++ src/services/users/get-like-first-name.js
@@ -11,16 +11,21 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose email matches the passed as parameter
- * @param {String} email String type
+ * @description get all paged users whose first name matches the passed as parameter
+ * @param {String} firstName String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
- * @param {Object} orderBy Array Object type
+ * @param {Object} order Array Object type
  * @returns a list of paginated users
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeEmail = async function (email, pageSizeNro, pageNro, orderBy) {
+const getLikeFirstName = async function (
+  firstName,
+  pageSizeNro,
+  pageNro,
+  order
+) {
   try {
     usersList = null;
     msg = null;
@@ -34,19 +39,19 @@ const getLikeEmail = async function (email, pageSizeNro, pageNro, orderBy) {
           ],
         },
         where: {
-          email: {
-            [Op.like]: `%${email}%`, //containing what is entered, less strictmatch
+          first_name: {
+            [Op.like]: `%${firstName}%`, //containing what is entered, less strictmatch
           },
         },
         limit: pageSizeNro,
         offset: pageNro,
-        order: orderBy,
+        order: order,
       })
         .then((users) => {
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeEmail User model. Caused by ${error}`;
+          msg = `Error in getLikeFirstName User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -54,15 +59,14 @@ const getLikeEmail = async function (email, pageSizeNro, pageNro, orderBy) {
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeEmail function. Caused by ${error}`;
+    msg = `Error in getLikeFirstName function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
-
   console.log(usersList);
   return usersList;
 };
 
 module.exports = {
-  getLikeEmail,
+  getLikeFirstName,
 };
