+++ src/services/users/get-like-identification-number.js
@@ -1,5 +1,5 @@
 //Externals
-const { Op, Sequelize } = require("sequelize");
+const { Op } = require("sequelize");
 //Models
 const { User } = require("../../models/sequelize/user");
 //Enums
@@ -11,23 +11,24 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose update date matches the passed as parameter
- * @param {String} updateDate String type
+ * @description get all paged users whose identification number matches the passed as parameter
+ * @param {String} identificationNumber String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
- * @param {Object} orderBy Array Object type
+ * @param {Object} order Array Object type
  * @returns a list of paginated users
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeUpdateDate = async function (
-  updateDate,
+const getLikeIdentificationNumber = async function (
+  identificationNumber,
   pageSizeNro,
   pageNro,
-  orderBy
+  order
 ) {
   try {
     usersList = null;
+    msg = null;
 
     if (User != null) {
       await User.findAll({
@@ -38,25 +39,19 @@ const getLikeUpdateDate = async function (
           ],
         },
         where: {
-          [Op.and]: [
-            //This case is for DATEONLY format
-            Sequelize.where(
-              Sequelize.fn("DATE", Sequelize.col("update_date")),
-              {
-                [Op.eq]: updateDate,
-              }
-            ),
-          ],
+          identification_number: {
+            [Op.like]: `%${identificationNumber}%`, //containing what is entered, less strictmatch
+          },
         },
         limit: pageSizeNro,
         offset: pageNro,
-        order: orderBy,
+        order: order,
       })
         .then((users) => {
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeUpdateDate User model. Caused by ${error}`;
+          msg = `Error in getLikeIdentificationNumber User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -64,7 +59,7 @@ const getLikeUpdateDate = async function (
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeUpdateDate function. Caused by ${error}`;
+    msg = `Error in getLikeIdentificationNumber function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
@@ -73,5 +68,5 @@ const getLikeUpdateDate = async function (
 };
 
 module.exports = {
-  getLikeUpdateDate,
+  getLikeIdentificationNumber,
 };
