+++ src/services/users/get-like-identification-type.js
@@ -11,20 +11,20 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose identification number matches the passed as parameter
- * @param {String} identificationNumber String type
+ * @description get all paged users whose identification type matches the passed as parameter
+ * @param {String} identificationType String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
- * @param {Object} orderBy Array Object type
+ * @param {Object} order Array Object type
  * @returns a list of paginated users
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeIdentificationNumber = async function (
-  identificationNumber,
+const getLikeIdentificationType = async function (
+  identificationType,
   pageSizeNro,
   pageNro,
-  orderBy
+  order
 ) {
   try {
     usersList = null;
@@ -39,19 +39,19 @@ const getLikeIdentificationNumber = async function (
           ],
         },
         where: {
-          identification_number: {
-            [Op.like]: `%${identificationNumber}%`, //containing what is entered, less strictmatch
+          identification_type: {
+            [Op.like]: `%${identificationType}%`, //containing what is entered, less strictmatch
           },
         },
         limit: pageSizeNro,
         offset: pageNro,
-        order: orderBy,
+        order: order,
       })
         .then((users) => {
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeIdentificationNumber User model. Caused by ${error}`;
+          msg = `Error in getLikeIdentificationType User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -59,7 +59,7 @@ const getLikeIdentificationNumber = async function (
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeIdentificationNumber function. Caused by ${error}`;
+    msg = `Error in getLikeIdentificationType function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
@@ -68,5 +68,5 @@ const getLikeIdentificationNumber = async function (
 };
 
 module.exports = {
-  getLikeIdentificationNumber,
+  getLikeIdentificationType,
 };
