+++ src/services/users/get-like-last-name.js
@@ -1,5 +1,5 @@
 //Externals
-const { Op } = require("sequelize");
+const { Sequelize, Op } = require("sequelize");
 //Models
 const { User } = require("../../models/sequelize/user");
 //Enums
@@ -11,24 +11,23 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose first name matches the passed as parameter
- * @param {String} firstName String type
+ * @description get all paged users whose last name matches the passed as parameter
+ * @param {String} lastname String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
- * @param {Object} orderBy Array Object type
+ * @param {Object} order Array Object type
  * @returns a list of paginated users
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeFirstName = async function (
-  firstName,
+const getLikeLastName = async function (
+  lastName,
   pageSizeNro,
   pageNro,
-  orderBy
+  order
 ) {
   try {
     usersList = null;
-    msg = null;
 
     if (User != null) {
       await User.findAll({
@@ -39,19 +38,19 @@ const getLikeFirstName = async function (
           ],
         },
         where: {
-          first_name: {
-            [Op.like]: `%${firstName}%`, //containing what is entered, less strictmatch
+          last_name: {
+            [Op.like]: `%${lastName}%`, //containing what is entered, less strictmatch
           },
         },
         limit: pageSizeNro,
         offset: pageNro,
-        order: orderBy,
+        order: order,
       })
         .then((users) => {
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeFirstName User model. Caused by ${error}`;
+          msg = `Error in getLikeLastName User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -59,7 +58,7 @@ const getLikeFirstName = async function (
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeFirstName function. Caused by ${error}`;
+    msg = `Error in getLikeLastName function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
@@ -68,5 +67,5 @@ const getLikeFirstName = async function (
 };
 
 module.exports = {
-  getLikeFirstName,
+  getLikeLastName,
 };
