+++ src/services/users/get-like-nickname.js
@@ -1,5 +1,5 @@
 //Externals
-const { Sequelize, Op } = require("sequelize");
+const { Op } = require("sequelize");
 //Models
 const { User } = require("../../models/sequelize/user");
 //Enums
@@ -11,20 +11,20 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose last name matches the passed as parameter
- * @param {String} lastname String type
+ * @description get all paged users whose nickName matches the passed as parameter
+ * @param {String} nickName String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
- * @param {Object} orderBy Array Object type
+ * @param {Object} order Array Object type
  * @returns a list of paginated users
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeLastName = async function (
-  lastName,
+const getLikeNickname = async function (
+  nickName,
   pageSizeNro,
   pageNro,
-  orderBy
+  order
 ) {
   try {
     usersList = null;
@@ -38,19 +38,19 @@ const getLikeLastName = async function (
           ],
         },
         where: {
-          last_name: {
-            [Op.like]: `%${lastName}%`, //containing what is entered, less strictmatch
+          nickname: {
+            [Op.like]: `%${nickName}%`, //containing what is entered, less strictmatch
           },
         },
         limit: pageSizeNro,
         offset: pageNro,
-        order: orderBy,
+        order: order,
       })
         .then((users) => {
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeLastName User model. Caused by ${error}`;
+          msg = `Error in getLikeNickname User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -58,7 +58,7 @@ const getLikeLastName = async function (
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeLastName function. Caused by ${error}`;
+    msg = `Error in getLikeNickname function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
@@ -67,5 +67,5 @@ const getLikeLastName = async function (
 };
 
 module.exports = {
-  getLikeLastName,
+  getLikeNickname,
 };
