+++ src/services/users/get-like-update-date.js
@@ -1,9 +1,9 @@
 //Externals
-const { Op } = require("sequelize");
+const { Op, Sequelize } = require("sequelize");
 //Models
 const { User } = require("../../models/sequelize/user");
 //Enums
-const { statusName } = require("../../enums/connection/statusName");
+const { statusName } = require("../../enums/connection/status-name");
 //Helpers
 const { getDateFormat } = require("../../helpers/sequelize/format/dateFormat");
 //Const/Vars
@@ -11,8 +11,8 @@ let usersList;
 let msg;
 
 /**
- * @description get all paged users whose nickName matches the passed as parameter
- * @param {String} nickName String type
+ * @description get all paged users whose update date matches the passed as parameter
+ * @param {String} updateDate String type
  * @param {Number} pageSizeNro Number type
  * @param {Number} pageNro Number type
  * @param {Object} orderBy Array Object type
@@ -20,8 +20,8 @@ let msg;
  * @example
  * [{"id":1,"nickname":"RAFA-CON","first_name":"Rafael","last_name":"Castro","email":"rafael_castro88@gmail.com","identification_type":"DNI","identification_number":"445938822","country_id":"AR","creation_date":"2023-02-12 21:18:11","update_date":"2023-02-12 21:18:11"},{"id".....]
  */
-const getLikeNickname = async function (
-  nickName,
+const getLikeUpdateDate = async function (
+  updateDate,
   pageSizeNro,
   pageNro,
   orderBy
@@ -38,9 +38,15 @@ const getLikeNickname = async function (
           ],
         },
         where: {
-          nickname: {
-            [Op.like]: `%${nickName}%`, //containing what is entered, less strictmatch
-          },
+          [Op.and]: [
+            //This case is for DATEONLY format
+            Sequelize.where(
+              Sequelize.fn("DATE", Sequelize.col("update_date")),
+              {
+                [Op.eq]: updateDate,
+              }
+            ),
+          ],
         },
         limit: pageSizeNro,
         offset: pageNro,
@@ -50,7 +56,7 @@ const getLikeNickname = async function (
           usersList = users;
         })
         .catch((error) => {
-          msg = `Error in getLikeNickname User model. Caused by ${error}`;
+          msg = `Error in getLikeUpdateDate User model. Caused by ${error}`;
           console.error(`${msg}. Stack error type : ${error.stack}`);
           usersList = statusName.CONNECTION_ERROR;
         });
@@ -58,7 +64,7 @@ const getLikeNickname = async function (
       usersList = statusName.CONNECTION_REFUSED;
     }
   } catch (error) {
-    msg = `Error in getLikeNickname function. Caused by ${error}`;
+    msg = `Error in getLikeUpdateDate function. Caused by ${error}`;
     console.error(`${msg}. Stack error type : ${error.stack}`);
     usersList = statusName.CONNECTION_ERROR;
   }
@@ -67,5 +73,5 @@ const getLikeNickname = async function (
 };
 
 module.exports = {
-  getLikeNickname,
+  getLikeUpdateDate,
 };
